requirements.txt
discord.py
py-dactyl
discord
requests
ruamel.yaml
aiohttp
mariadb
aioconsole